Fix: Incorrect elements removed on "Mark all comments read"
When the user clicked "Mark all comments read", the script was updating
the wrong <tr/> element, presumably due to recent changes to the
comments page structure.  This caused most of the page (not just
comments) to disappear.

This commit updates the script to find each comment's <tr/> by looking
up the "comment_tr_NNN" class name instead of attempting to find the
element based on a series of fixed .parentNode lookups.  This should
provide more resiliency for future changes.