Fix to restore "show parent" feature
The "show parent" feature hasn't worked for some time, due to some DOM
restructuring of the page leaving the script unable to find the <span
class="comhead"/> element.  This fix uses a jQuery selector to find the
comhead element, no matter where it is in the comment's branch of the
DOM tree.